+++ CHANGELOG.md
@@ -3,6 +3,10 @@ SPDX-FileCopyrightText: 2020 Robin Krahl <robin.krahl@ireas.org>
 SPDX-License-Identifier: CC0-1.0
 -->
 
+# Unreleased
+
+- Add `style_mut` method to `StyledStr` and `StyledString`.
+
 # v0.1.0 (2020-09-24)
 
 Initial release with support for `ansi_term`, `crossterm`, `cursive`, `syntect`